requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.128.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
playwright>=1.40.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
supabase>=2.0.0
langchain==0.3.25
langchain-core==0.3.64
langchain-openai==0.3.21
langchain-community==0.3.24